add: add -D/--ignore-dep-versions
This flag is useful when using poudriere-bulk with the -S flag,
which skips recursive rebuilds of all dependents of a package.

In that situation, when building a port, Mk/Scripts/do-depends.sh would use
pkg-add to install a dependency (1), and if one of *its* dependencies (2) has
been updated, pkg-add wouldn't be able to find it, as the manifest of (1)
references an older version of (2) still, because we did not rebuild (1).
And the exact version lookup would fail.